// src/models/sofa-type.model.ts
export enum SofaType {
  DIVANO_3_PL_MAXI = 'divano_3_pl_maxi',
  DIVANO_3_PL = 'divano_3_pl',
  DIVANO_2_PL = 'divano_2_pl',
  CHAISE_LONGUE = 'chaise_longue',
  POUF_50_X_50 = 'pouf_50_x_50',
  POUF_60_X_60 = 'pouf_60_x_60',
  POUF_70_X_70 = 'pouf_70_x_70',
  ELEMENTO_SENZA_BRACCIOLO = 'elemento_senza_bracciolo',
  ELEMENTO_CON_BRACCIOLO = 'elemento_con_bracciolo',
  POLTRONA_90_CM = 'poltrona_90_cm',
  POLTRONA_80_CM = 'poltrona_80_cm',
  POLTRONA_70_CM = 'poltrona_70_cm',
  CUSTOM = 'custom',
}